+++ src/stac_pydantic_extensions/extensions/view.py
@@ -4,11 +4,13 @@ from typing import TYPE_CHECKING, ClassVar, Literal
 
 from pydantic import AnyUrl, ConfigDict
 
+from stac_pydantic_extensions import Collection
 from stac_pydantic_extensions.extensions._base import (
     BaseExtension,
     BaseExtraFields,
     MaturityLevel,
     OldBaseExtension,
+    as_summary_fields,
     prefix_alias,
 )
 from stac_pydantic_extensions.model_annotations import Azimuth, Elevation, OffNadir
@@ -114,6 +116,8 @@ class ViewGeometryExtension(BaseExtension):
         )
 
         model = FIELD_MODELS[stac_ext_version]
+        if isinstance(stac_object, Collection):
+            model = as_summary_fields(model)
         fields = model.model_validate(properties or {})
 
         if migrate and stac_ext_version != cls.version: